Now ASCII image with density String
you can now input your own density string to use instead of 'e' and here I'm using the {brightness} variable to set the size of size of the char. by also I'm also using it to map({brightness) between 0 and length of the density String. so the brighter the pixel the higher up the density string the char will be. or the other way around.